Created Ex Patch branch
This branch is for building a framework with various hooks for custom sprite and level code.

Added 3 hooks for level code: pre-NMI, post-NMI, and post-logic.
There is a table spanning all level ids 00-FF for each of these hooks! By default they use a null routine that is a simple RTS. These tables give hackers the ability to inject custom code in 3 separate places within the level code. Code is level specific so you don't have to worry about levels that use the same NMI or logic behaviors anymore.