Added RAM labels
Added a lot of labels to RAM and replaced all known RAM references with labels.
Added 2 level struct variants, one for alternate level RAM and another for level RAM that is referenced via index register offsets. There may be a better way to do this in the future but it's better than nothing for now.